requirements.txt
numpy==1.21.0
pyelftools==0.27